chore: split plugins
